From tomprince: both builds and steps now have getProperty and getProperties.
You can get at the properties object by calling getProperties in any case,
which is all WithProperty needs, so this doesn't break the mail notifiers.

The getProperties/render could probably be dropped from
Properties. But that would mean changing the tests.  It would need
some paramater renaming, and documenting that the lambdas passed to
WithProperties get one of a BuildStep, Build, or Properties instance.

https://gist.github.com/973698